Release 1.3.5:
Theme manager now keeps track of the current active theme and alerts the components that need to know of a theme change via an event.

Users can now choose a board size from 3-6.

UI formatting tweaks to make it more dynamic.

Enabled Coin Toss again. Not sure when it was disabled.